Refactor, create view model directory
