+++ il_app/lib/logic/vm/entry_page_view_model.dart
