+++ il_app/lib/ui/pages/entry_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:il_app/logic/entry_page_view_model.dart';
+import 'package:il_app/logic/vm/entry_page_view_model.dart';
 import 'package:il_app/ui/widgets/message_card.dart';
 import 'package:provider/provider.dart';
